background: Scale by fractional output scale

Closes #338

See merge request World/Phosh/phosh!855

+++ src/background.h
@@ -18,7 +18,7 @@ G_DECLARE_FINAL_TYPE (PhoshBackground, phosh_background, PHOSH, BACKGROUND, Phos
 
 GtkWidget *phosh_background_new (gpointer layer_shell,
                                  gpointer wl_output,
-                                 guint scale,
+                                 float scale,
                                  gboolean primary);
 void phosh_background_set_primary (PhoshBackground *self, gboolean primary);
-void phosh_background_set_scale (PhoshBackground *self, guint scale);
+void phosh_background_set_scale (PhoshBackground *self, float scale);
